scripts/test_projection_methods.py: fix dtype mismatch in projection comparison
The activation tensors collected from float16 models caused
dtype mismatches when multiplied with float32 projection
matrices. This resulted in RuntimeError during the projection
comparison phase.

Convert activations to float32 before projection operations to
ensure consistent dtypes throughout the comparison. The
projection quality metrics are computed in float32 regardless
of the original model dtype.

Generated-by: Claude AI

+++ scripts/test_projection_methods.py
@@ -207,9 +207,10 @@ def compare_projection_methods(
     print(f"Compression ratio: {compression_ratio} ({d_compressed}/{d_in})")
 
     # Test data (held out from training)
+    # Convert to float32 for projection operations
     test_size = min(500, activations.shape[0] // 4)
-    test_data = activations[-test_size:].to(device)
-    train_data = activations[:-test_size]
+    test_data = activations[-test_size:].to(device).float()
+    train_data = activations[:-test_size].float()
 
     # Method 1: Random Orthogonal
     print(f"\n{'=' * 80}")
